+++ packages/cmf/src/reduxstorage/index.js
@@ -1,9 +1,0 @@
-/**
- * This is a CMF plugin that let you configure your store
- */
-
-import * as reduxLocalStorage from './reduxLocalStorage';
-
-export default {
-	localStorage: reduxLocalStorage,
-};