+++ packages/forms/src/UIForm/utils/labels.js
@@ -1,6 +1,10 @@
+import classnames from 'classnames';
+
 import { ButtonIcon, Popover, StackHorizontal } from '@talend/design-system';
 
-export const getLabelProps = (title, labelProps, hint) => {
+import styles from './labels.module.scss';
+
+export const getLabelProps = (title, labelProps, hint, required) => {
 	if (!hint) {
 		return {
 			children: title,
@@ -9,8 +13,8 @@ export const getLabelProps = (title, labelProps, hint) => {
 	}
 	return {
 		children: (
-			<StackHorizontal gap="XS" align="center">
-				{title}
+			<StackHorizontal gap="XXS" align="center">
+				<span className={classnames({ [styles.required]: required })}>{title}</span>
 				<Popover
 					position={hint.overlayPlacement || 'auto'}
 					data-test={hint['data-test']}
@@ -28,5 +32,6 @@ export const getLabelProps = (title, labelProps, hint) => {
 			</StackHorizontal>
 		),
 		...labelProps,
+		required: false,
 	};
 };