+++ packages/playground-vite/mockVite/mock/kit/index.js
@@ -1,0 +1,8 @@
+/* eslint-disable global-require */
+import example from './example.json';
+import validation from './validation.json';
+
+export default {
+	example,
+	validation,
+};