+++ fork/json-schema-form-core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @talend/json-schema-form-core
 
+## 1.4.1
+
+### Patch Changes
+
+- f321a0d: Remove unused tsconfig.esm.json (initially added to use TSC but we stay with babel at the end)
+
+## 1.4.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+## 1.3.0
+
+### Minor Changes
+
+- 3bd16fc: Add support to ESM
+
 ## 1.2.7
 
 ### Patch Changes