+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @talend/icons
 
+## 7.10.2
+
+### Patch Changes
+
+- f321a0d: Remove unused tsconfig.esm.json (initially added to use TSC but we stay with babel at the end)
+
+## 7.10.1
+
+### Patch Changes
+
+- b8f7672: fix module not found error for @talend/react-bootstrap and @talend/icons
+
+## 7.10.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+## 7.9.0
+
+### Minor Changes
+
+- 3bd16fc: Add support to ESM
+
+### Patch Changes
+
+- d053412: Fix esm script
+- 9b66a09: Improve changes needed for cypress and simplify scripts
+
+## 7.8.0
+
+### Minor Changes
+
+- d409b57: feat(TMC-27713/icons): update qlik favicon
+
 ## 7.7.2
 
 ### Patch Changes