+++ packages/router-bridge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @talend/router-bridge
 
+## 2.2.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
 ## 2.1.2
 
 ### Patch Changes