+++ packages/storybook-one/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @talend/ui-storybook-one
 
+## 2.5.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/react-faceted-search@20.0.0
+  - @talend/design-system@11.0.0
+  - @talend/react-components@17.0.0
+  - @talend/react-dataviz@7.0.0
+  - @talend/react-forms@15.0.0
+  - @talend/storybook-docs@2.5.0
+  - @talend/design-tokens@3.4.0
+  - @talend/icons@7.10.0
+
 ## 2.4.4
 
 ### Patch Changes