+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @talend/utils
 
+## 3.2.3
+
+### Patch Changes
+
+- 7ebe036: Revert "fix: Fix exports to allow deep imports"
+
+## 3.2.2
+
+### Patch Changes
+
+- f321a0d: Remove unused tsconfig.esm.json (initially added to use TSC but we stay with babel at the end)
+
+## 3.2.1
+
+### Patch Changes
+
+- 94aad7d: fix: Fix exports to allow deep imports
+
+## 3.2.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+## 3.1.0
+
+### Minor Changes
+
+- 3bd16fc: Add support to ESM
+
 ## 3.0.4
 
 ### Patch Changes