+++ tools/scripts-config-prettier/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @talend/scripts-config-prettier
 
+## 12.2.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
 ## 12.1.3
 
 ### Patch Changes