+++ tools/scripts-locales/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @talend/scripts-locales
 
+## 1.2.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/scripts-utils@2.3.0
+
 ## 1.1.2
 
 ### Patch Changes