+++ tools/scripts-publish-local/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @talend/scripts-publish-local
 
+## 1.4.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
 ## 1.3.3
 
 ### Patch Changes